Physics 1

+++ src/Eng/Scene/Comp/LightProbe.cpp
@@ -35,15 +35,15 @@ void LightProbe::Write(const LightProbe &pr, JsObjectP &js_out) {
     { // write offset
         JsArrayP js_offset(alloc);
 
-        js_offset.Push(JsNumber(double(pr.offset[0])));
-        js_offset.Push(JsNumber(double(pr.offset[1])));
-        js_offset.Push(JsNumber(double(pr.offset[2])));
+        js_offset.Push(JsNumber(pr.offset[0]));
+        js_offset.Push(JsNumber(pr.offset[1]));
+        js_offset.Push(JsNumber(pr.offset[2]));
 
         js_out.Push("offset", std::move(js_offset));
     }
 
     { // write radius
-        js_out.Push("radius", JsNumber(double(pr.radius)));
+        js_out.Push("radius", JsNumber(pr.radius));
     }
 
     { // write sh coefficients
@@ -52,9 +52,9 @@ void LightProbe::Write(const LightProbe &pr, JsObjectP &js_out) {
         for (int i = 0; i < 4; i++) {
             JsArrayP js_coeff(alloc);
 
-            js_coeff.Push(JsNumber(double(pr.sh_coeffs[i][0])));
-            js_coeff.Push(JsNumber(double(pr.sh_coeffs[i][1])));
-            js_coeff.Push(JsNumber(double(pr.sh_coeffs[i][2])));
+            js_coeff.Push(JsNumber(pr.sh_coeffs[i][0]));
+            js_coeff.Push(JsNumber(pr.sh_coeffs[i][1]));
+            js_coeff.Push(JsNumber(pr.sh_coeffs[i][2]));
 
             js_coeffs.Push(std::move(js_coeff));
         }
